Moving adapter for bn256.
