+++ suites/all.go
@@ -3,7 +3,6 @@ package suites
 import (
 	"go.dedis.ch/kyber/v3/group/edwards25519"
 	"go.dedis.ch/kyber/v3/group/nist"
-	"go.dedis.ch/kyber/v3/pairing"
 	"go.dedis.ch/kyber/v3/pairing/bn256"
 )
 
@@ -15,7 +14,7 @@ func init() {
 	register(bn256.NewSuiteG1())
 	register(bn256.NewSuiteG2())
 	register(bn256.NewSuiteGT())
-	register(pairing.NewSuiteBn256())
+	register(bn256.NewSuiteBn256())
 	// This is a constant time implementation that should be
 	// used as much as possible
 	register(edwards25519.NewBlakeSHA256Ed25519())
